add controllers for workout programs
