+++ backend/routes/workout_programs.js
@@ -1,96 +1,22 @@
 const express = require('express');
+const { getPrograms,createPrograms,getSpecificProgram,updateProgram, deleteProgram } = require('../controllers/workoutPrograms');
 const workout_programs = require('../models/workout_programs');
 
 const router = express.Router();
 
-//get post
-router.get('/programs', (req, res) => {
-    workout_programs.find().exec((err,programs) => {
-        if(err){
-            return res.status(400).json({
-                error: err,
-                success: false
-            })
-        } else {
-            return res.status(200).json({
-                success: true,
-                existingPrograms: programs
-            })
-        }
-    })
-});
+//get programs
+router.get('/programs', getPrograms);
 
 //save new program
-router.post('/program/save', (req, res) => {
-    let newProgram = workout_programs(req.body);
-    newProgram.save((err) => {
-        if(err){
-            return res.status(400).json({
-                error: err,
-                success: false
-            });
-        } else {
-            return res.status(200).json({
-                success: true
-            })
-        }
-
-    });
-});
+router.post('/program/save', createPrograms);
 
 //get a specific program
-router.get('/program/:id', (req, res) => {
-    let programId = req.params.id;
-
-    workout_programs.findById((programId), (err, program) => {
-        if(err){
-            return res.status(400).json({success: false, err})
-        } else {
-            return res.status(200).json({
-                success: true,
-                program: program
-            })
-        }
-    })
-});
+router.get('/program/:id', getSpecificProgram);
 
 //update program
-router.put('/program/update/:id', (req, res) => {
-    workout_programs.findByIdAndUpdate(
-        req.params.id,
-        {
-            $set:req.body
-        },
-        (err,program) => {
-            if (err){
-                return res.status(400).json({
-                    error: err
-                })
-            } else {
-                return res.status(200).json({
-                    seccess: "updated successfully"
-                })
-            }
-        }
-    )
-});
+router.put('/program/update/:id', updateProgram);
 
 //delete
-router.delete('/program/delete/:id', (req, res) => {
-    workout_programs.findByIdAndRemove(req.params.id).exec((err,deletedProgram) => {
-        if (err){
-            return res.status(400).json({
-                error: err
-            })
-        } else { 
-            return res.status(200).json({
-                seccess: "deleted successfully", deletedProgram
-            })
-        }
-    })
-});
-
-
-
+router.delete('/program/delete/:id', deleteProgram);
 
 module.exports = router;
